# [processor/lsminterval] Define cardinality limits and handle overflows