--- a/processor/lsmintervalprocessor/internal/merger/key.go
+++ b/processor/lsmintervalprocessor/internal/merger/key.go
@@ -1,0 +1,76 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package merger // import "github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger"
+
+import (
+	"encoding/binary"
+	"errors"
+	"time"
+)
+
+// TODO (lahsivjar): Think about multitenancy, should be part of the key
+type Key struct {
+	Interval       time.Duration
+	ProcessingTime time.Time
+}
+
+// NewKey creates a new instance of the merger key.
+func NewKey(ivl time.Duration, pTime time.Time) Key {
+	return Key{
+		Interval:       ivl,
+		ProcessingTime: pTime,
+	}
+}
+
+// SizeBinary returns the size of the Key when binary encoded.
+// The interval, represented by time.Duration, is encoded to
+// 2 bytes by converting it into seconds. This allows a max of
+// ~18 hours duration.
+func (k *Key) SizeBinary() int {
+	// 2 bytes for interval, 8 bytes for processing time
+	return 10
+}
+
+// Marshal marshals the key into binary representation.
+func (k *Key) Marshal() ([]byte, error) {
+	ivlSeconds := uint16(k.Interval.Seconds())
+
+	var (
+		offset int
+		d      [10]byte
+	)
+	binary.BigEndian.PutUint16(d[offset:], ivlSeconds)
+	offset += 2
+
+	binary.BigEndian.PutUint64(d[offset:], uint64(k.ProcessingTime.Unix()))
+
+	return d[:], nil
+}
+
+// Unmarshal unmarshals the binary representation of the Key.
+func (k *Key) Unmarshal(d []byte) error {
+	if len(d) != 10 {
+		return errors.New("failed to unmarshal key, invalid sized buffer provided")
+	}
+	var offset int
+	k.Interval = time.Duration(binary.BigEndian.Uint16(d[offset:2])) * time.Second
+	offset += 2
+
+	k.ProcessingTime = time.Unix(int64(binary.BigEndian.Uint64(d[offset:offset+8])), 0)
+	return nil
+}