--- a/processor/lsmintervalprocessor/testdata/exphistogram_cumulative_overflow/next.yaml
+++ b/processor/lsmintervalprocessor/testdata/exphistogram_cumulative_overflow/next.yaml
@@ -1,0 +1,1 @@
+resourceMetrics: []