--- a/processor/lsmintervalprocessor/testdata/histogram_delta_overflow/next.yaml
+++ b/processor/lsmintervalprocessor/testdata/histogram_delta_overflow/next.yaml
@@ -1,0 +1,1 @@
+resourceMetrics: []