Related to: https://github.com/elastic/opentelemetry-collector-components/issues/141

For more details checkout https://github.com/elastic/opentelemetry-collector-components/issues/141#issuecomment-2508751134 comment.