@lahsivjar 2024-12-10
@axw I have addressed all of the comments other than one and I am marking this ready for review now. The performance consideration is still there and I have a few threads to chase down for improvements but I don't have a lot of hope from them, so, if you have any ideas (even rough) I will be more than happy to chase them down.

I have left adding metrics limits to a follow-up PR. I will also add a few more detailed overflow tests.