@lahsivjar 2024-12-19
Update so far, I have been able to make some progress on optimizing the processor as a whole. I need to perform more tests  and more real-world correctness tests but with the changes so far below are the result as compared to main branch (TL;DR around 50% for allocations and runtime).

<details>
<summary> `BenchmarkAggregation` as compared to `main` branch </summary>

name                                            old time/op    new time/op    delta
Aggregation/sum_cumulative-10                     10.3µs ± 5%     5.3µs ± 2%  -48.24%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                          10.4µs ± 2%     5.5µs ± 3%  -47.77%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10               10.8µs ± 2%     5.8µs ± 4%  -45.96%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                    11.2µs ± 5%     6.1µs ± 8%  -45.48%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10            12.0µs ± 4%     6.2µs ± 3%  -47.84%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                 11.8µs ± 3%     6.6µs ± 6%  -44.36%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                    10.9µs ± 3%     5.9µs ± 1%  -46.36%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                1.35µs ± 2%    1.33µs ± 3%     ~     (p=0.341 n=5+5)
AggregationWithOTTL/sum_cumulative-10             12.5µs ± 3%     5.5µs ± 2%  -56.30%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                  12.8µs ± 2%     5.5µs ± 2%  -57.41%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10       13.1µs ± 2%     6.0µs ± 3%  -53.86%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10            13.4µs ± 2%     6.2µs ± 3%  -53.69%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10    13.3µs ± 3%     6.2µs ± 6%  -53.25%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta-10         13.4µs ± 2%     6.6µs ± 0%  -50.77%  (p=0.016 n=5+4)
AggregationWithOTTL/summary_enabled-10            12.7µs ± 1%     6.2µs ± 3%  -51.40%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10        1.34µs ± 2%    1.36µs ± 7%     ~     (p=0.500 n=5+5)

name                                            old alloc/op   new alloc/op   delta
Aggregation/sum_cumulative-10                     17.5kB ± 1%     8.6kB ± 0%  -51.06%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                          17.6kB ± 1%     8.6kB ± 2%  -51.17%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10               19.5kB ± 1%    10.2kB ± 1%  -47.56%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                    19.4kB ± 1%    10.1kB ± 0%  -47.81%  (p=0.016 n=5+4)
Aggregation/exphistogram_cumulative-10            21.7kB ± 2%    10.1kB ± 0%  -53.39%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                 22.2kB ± 1%    10.6kB ± 2%  -52.34%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                    18.3kB ± 1%     8.8kB ± 0%  -51.79%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                1.37kB ± 0%    1.32kB ± 0%   -3.99%  (p=0.016 n=4+5)
AggregationWithOTTL/sum_cumulative-10             20.2kB ± 1%     8.6kB ± 1%  -57.46%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                  20.3kB ± 1%     8.6kB ± 3%  -57.72%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10       22.1kB ± 3%    10.3kB ± 1%  -53.55%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10            22.1kB ± 2%    10.2kB ± 1%  -54.09%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10    23.0kB ± 4%    10.2kB ± 2%  -55.62%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta-10         23.0kB ± 1%    10.6kB ± 0%  -53.80%  (p=0.016 n=5+4)
AggregationWithOTTL/summary_enabled-10            20.8kB ± 2%     9.0kB ± 2%  -56.84%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10        1.38kB ± 0%    1.32kB ± 0%   -4.22%  (p=0.008 n=5+5)

name                                            old allocs/op  new allocs/op  delta
Aggregation/sum_cumulative-10                        199 ± 1%       107 ± 0%  -46.23%  (p=0.016 n=5+4)
Aggregation/sum_delta-10                             200 ± 0%       111 ± 5%  -44.61%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10                  212 ± 0%       121 ± 4%  -43.02%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                       214 ± 0%       125 ± 0%  -41.48%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10               228 ± 0%       120 ± 1%  -47.60%  (p=0.016 n=4+5)
Aggregation/exphistogram_delta-10                    237 ± 0%       131 ± 6%  -44.81%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                       212 ± 0%       114 ± 1%  -46.23%  (p=0.016 n=4+5)
Aggregation/summary_passthrough-10                  37.0 ± 0%      33.0 ± 0%  -10.81%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_cumulative-10                227 ± 0%       109 ± 3%  -51.89%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                     229 ± 0%       109 ± 7%  -52.40%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10          244 ± 0%       124 ± 0%  -49.10%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10               245 ± 0%       123 ± 3%  -49.96%  (p=0.016 n=4+5)
AggregationWithOTTL/exphistogram_cumulative-10       256 ± 0%       124 ± 4%  -51.72%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_delta-10            265 ± 0%       138 ± 0%  -47.92%  (p=0.029 n=4+4)
AggregationWithOTTL/summary_enabled-10               240 ± 0%       120 ± 6%  -50.17%  (p=0.008 n=5+5)
AggregationWithOTTL/summary_passthrough-10          37.0 ± 0%      33.0 ± 0%  -10.81%  (p=0.008 n=5+5)

</details>

I am working on a few more ideas for optimizations, I hope to bring the allocations and the time/op down by at least 10% 🤞