@lahsivjar 2024-12-19
(deleted the above comment on performance improvements)

I have made a silly mistake in the `Merge` function implementation due to which the recent commits are in an incorrect state 🤦 .